/ Exact calculation of relative humidity
Calculating the relative humidity back from the dewpoint measurements.
For the creation of BUFR files, the measured RH is converted to
dewpoint temperature internally by the sounding unit (here MW41).
The inverse formula has now been applied to retrieve the originally
measured RHs.

This has an effect on the output of the relative humidity.

Fixes #3